// src/constants/types/roles.ts
export enum Role {
  Aida = '@Aida',
  Amneris = '@Amneris',
  Radames = '@Radames',
  Amonasro = '@Amonasro',
  Ramfis = '@Ramfis',
  IlRe = '@Il Re',
  UnMessagero = '@Il Messagero',
  HighPriestess = '@High Priestess',
  Chorus = '@Chorus',
  MaleChorus = '@Male Chorus',
  FemaleChorus = '@Female Chorus',
  MaleDancers = '@Male Dancers',
  FemaleDancers = '@Female Dancers',
  Tutti = '@Tutti'
}